Move EmojiPickerViewModel
