+++ app/src/main/java/com/geeksville/mesh/ui/common/components/EmojiPicker.kt
@@ -29,9 +29,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.emoji2.emojipicker.RecentEmojiProviderAdapter
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
-import com.geeksville.mesh.ui.common.EmojiPickerViewModel
 import com.geeksville.mesh.util.CustomRecentEmojiProvider
 import org.meshtastic.core.ui.component.BottomSheetDialog
+import org.meshtastic.core.ui.emoji.EmojiPickerViewModel
 
 @Composable
 fun EmojiPicker(
